pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quarto-mssql-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "google-genai>=1.47.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "vl-convert-python>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]
